Move instrument.py to _instrument.py
This allows us to make clear that the internal location
of the LSST gen3 instrument classes is not relevant
to the user of the package.
